update style to use class, add convert bytes to human readble to utils.js
